update parsing diagnostics

+++ step4_parse/ParseAll.ps1
@@ -23,23 +23,25 @@ Get-ChildItem -Path $inDir -Filter "tokens_*.csv" | ForEach-Object {
     $errorFile = Join-Path -Path $errorDir -ChildPath (Split-Path -Leaf $errorFile)
 
     if ((-not (Test-Path $outFile)) -or $Force) {
-        try {
-            if (Test-Path $errorFile) {
-                Remove-Item -Path $errorFile -Force
-            }
-
-            . $PSScriptRoot\ParseFile.ps1 -InFile $inFile |
-                ConvertTo-Json -Depth 100 |
-                Set-Content -Path $outFile -Encoding UTF8
+        if (Test-Path $errorFile) {
+            Remove-Item -Path $errorFile -Force
         }
-        catch {
+
+        $parse = . $PSScriptRoot\ParseFile.ps1 -InFile $inFile
+        if ($parse.ParseOk) {
+            $parse | ConvertTo-Json -Depth 100 | Set-Content -Path $outFile -Encoding UTF8
+        } else {
             # delete the outFile if partially created
             if (Test-Path $outFile) {
                 Remove-Item $outFile
             }
 
-            $errorMessage = "Failed to parse $inFile`n$_"
-            $errorMessage += "`nStack Trace:`n$($_.ScriptStackTrace)"
+            $errorMessage = "Failed to parse $inFile"
+            $errorMessage += "`n`n"
+            $errorMessage += $parse.ParseDiagnostics | ConvertTo-Json -Depth 100
+            $errorMessage += "`n`n"
+            $errorMessage += $parse | ConvertTo-Json -Depth 100 
+            # $errorMessage += "`nStack Trace:`n$($_.ScriptStackTrace)"
 
             # Include input file contents for debugging
             $errorMessage += "`n`nInput File:`n"
